+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Unrealeased
+
+### New Features
+
+* OpenShell3D: project_coincident_faces_of
+*
+
+### Fixed
+
+*
+*
+
+### Performance improvements
+
+*
+*
+
+### Refactorings
+
+*
+*
+
+
 ## v0.5.0
 
 ### New Features